move `Argument` component to `@graphiql/react`
