+++ packages/graphiql-react/src/explorer/components/argument.css
@@ -1,0 +1,22 @@
+.graphiql-doc-explorer-argument {
+  & > * + * {
+    margin-top: var(--px-12);
+  }
+}
+
+.graphiql-doc-explorer-argument-name {
+  color: var(--color-purple);
+}
+
+.graphiql-doc-explorer-argument-deprecation {
+  background-color: var(--color-orche-background);
+  border: 1px solid var(--color-orche);
+  border-radius: var(--border-radius-4);
+  color: var(--color-orche);
+  padding: var(--px-8);
+}
+
+.graphiql-doc-explorer-argument-deprecation-label {
+  font-size: var(--font-size-hint);
+  font-weight: var(--font-weight-medium);
+}
