+++ packages/graphiql/__mocks__/@graphiql/react.tsx
@@ -12,6 +12,7 @@ import {
 import React, { useEffect, useRef, useState } from 'react';
 
 export {
+  Argument,
   ChevronDownIcon,
   ChevronUpIcon,
   CopyIcon,
